Feat: Response 분리 / 이동
- Paging을 위하여 response 분리

# KIT-65
